+++ docs/project/changelog.md
@@ -14,6 +14,12 @@ substitutions:
 
 ## Unreleased
 
+### Packages
+
+- New packages: sqlalchemy {pr}`2112`
+
+- Upgraded packages: distlib (0.3.4)
+
 ### Uncategorized
 
 - {{Fix}} A fatal error in `scipy.stats.binom.ppf` has been fixed.
@@ -21,6 +27,17 @@ substitutions:
 
 - {{Fix}} Type signature mismatches in some numpy comparators have been fixed.
   {pr}`2110`
+
+- {{Enhancement}} Pyodide now uses Python wheel files to distribute packages
+  rather than the emscripten `file_packager.py` format.
+  {pr}`2027`
+
+- {{Fix}} The "PyProxy has already been destroyed" error message has been
+  improved with some context information.
+  {pr}`2121`
+
+- {{Enhancement}} Pressing TAB in REPL no longer triggers completion when input
+  is whitespace. {pr}`2125`
 
 ## Version 0.19.0